# docs/swagger.yaml
definitions:
...
  data.Country:
    properties:
      active:
        type: boolean
      name:
        type: string
    type: object
...
  routes.MessageResponse:
    properties:
      code:
        type: integer
      details:
        type: string
    type: object
...
  routes.roleParams:
    properties:
      name:
        type: string
    type: object
...
  routes.userParams:
    properties:
      country:
        type: string
      description:
        type: string
      dni:
        type: string
      email:
        type: string
      name:
        type: string
      nickname:
        type: string
      password:
        type: string
    required:
    - country
    - dni
    - email
    - name
    - nickname
    type: object
...
  routes.userStruct:
    properties:
      active:
        type: boolean
      country:
        $ref: '#/definitions/data.Country'
      email:
        type: string
      id:
        type: string
      name:
        type: string
      nick_name:
        type: string
      picture:
        type: string
      roles:
        items:
          $ref: '#/definitions/routes.roleParams'
        type: array
      status_user:
        type: string
      validated:
        type: boolean
    type: object
...
info:
  contact: {}
  description: api autentia manager
  title: Autentia Manager Documentation
  version: "1"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.userStruct'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/routes.userParams'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario Creado
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: user
      tags:
      - user
...
securityDefinitions:
  BarerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"